Added git setting on/off feature

+++ tabstasher.py
@@ -24,17 +24,18 @@ class TabstashCommand(sublime_plugin.WindowCommand):
         default_stashes = default_settings.get('stashes')
         default_stashes.append(self.new_stash_obj)
         default_settings.set('stashes',default_stashes)
-        git_stash_cmd = 'git stash save tabstasher' + stash
-        process = subprocess.call( 'cd' + self.root + ';'+ git_stash_cmd, shell=True)
+        if default_settings.get('git'):
+            git_stash_cmd = 'git stash save tabstasher' + stash
+            process = subprocess.call( 'cd' + self.root + ';'+ git_stash_cmd, shell=True)
         self.window.run_command('close_all')
 
 class TabunstashCommand(sublime_plugin.WindowCommand):
 
     def run(self):
         self.root = self.window.folders()[0]
-        default_settings = sublime.load_settings("Tabstasher.sublime-settings")
-        if default_settings.has('stashes'):
-            self.array_of_stashes = default_settings.get('stashes')
+        self.default_settings = sublime.load_settings("Tabstasher.sublime-settings")
+        if self.default_settings.has('stashes'):
+            self.array_of_stashes = self.default_settings.get('stashes')
             array_of_stashes_names = []
             for stash in self.array_of_stashes:
                 array_of_stashes_names.append(stash['name'])
@@ -45,9 +46,9 @@ class TabunstashCommand(sublime_plugin.WindowCommand):
             self.window.run_command('close_all')
             for stashedFile in self.array_of_stashes[picked]['files']:
                 self.window.open_file(stashedFile)
-            git_apply_cmd = 'git stash apply --tabstasher' + self.array_of_stashes[picked]['name']
-            process = subprocess.call( 'cd' + self.root + ';'+ git_apply_cmd, shell=True)
-            process.communicate()[0]
+            if self.default_settings.get('git'):
+                git_apply_cmd = 'git stash apply --tabstasher' + self.array_of_stashes[picked]['name']
+                process = subprocess.call( 'cd' + self.root + ';'+ git_apply_cmd, shell=True)
 
 class TabpoplaststashCommand(sublime_plugin.WindowCommand):
 
@@ -59,8 +60,9 @@ class TabpoplaststashCommand(sublime_plugin.WindowCommand):
             array_of_stashes = default_settings.get('stashes')
             for stashedFile in array_of_stashes[-1]['files']:
                 self.window.open_file(stashedFile)
-            git_poplast_cmd = 'git stash pop --tabstasher' + array_of_stashes[-1]['name']
-            process = subprocess.call( 'cd' + self.root + ';'+ git_poplast_cmd, shell=True)
+            if default_settings.get('git'):
+                git_poplast_cmd = 'git stash pop --tabstasher' + array_of_stashes[-1]['name']
+                process = subprocess.call( 'cd' + self.root + ';'+ git_poplast_cmd, shell=True)
             del array_of_stashes[-1]
             default_settings.set('stashes',array_of_stashes)
 
@@ -82,7 +84,8 @@ class TabpopstashCommand(sublime_plugin.WindowCommand):
             for stashedFile in self.array_of_stashes[picked]['files']:
                 self.window.open_file(stashedFile)
             git_pop_cmd = 'git stash pop --tabstasher' + array_of_stashes[picked]['name']
-            process = subprocess.call( 'cd' + self.root + ';'+ git_pop_cmd, shell=True)
+            if self.default_settings.get('git'):
+                process = subprocess.call( 'cd' + self.root + ';'+ git_pop_cmd, shell=True)
             del self.array_of_stashes[picked]
             self.default_settings.set('stashes',self.array_of_stashes)
 
@@ -102,7 +105,8 @@ class TabdeletestashCommand(sublime_plugin.WindowCommand):
     def on_done(self, picked):
         if picked >= 0:
             git_drop_cmd = 'git stash drop --tabstasher' + self.array_of_stashes[picked]['name']
-            process = subprocess.call( 'cd' + self.root + ';'+ git_drop_cmd, shell=True)
+            if self.default_settings.get('git'):
+                process = subprocess.call( 'cd' + self.root + ';'+ git_drop_cmd, shell=True)
             del self.array_of_stashes[picked]
             self.default_settings.set('stashes',self.array_of_stashes)
 
@@ -113,7 +117,17 @@ class TabclearstashCommand(sublime_plugin.WindowCommand):
         self.root = self.window.folders()[0]
         if self.default_settings.has('stashes'):
             stashes = self.default_settings.get('stashes')
-            for stash in stashes:
-                git_drop_cmd = 'git stash drop --tabstasher' + stash['name']
-                process = subprocess.call( 'cd' + self.root + ';'+ git_drop_cmd, shell=True)
+            if self.default_settings.get('git'):
+                for stash in stashes:
+                    git_drop_cmd = 'git stash drop --tabstasher' + stash['name']
+                    process = subprocess.call( 'cd' + self.root + ';'+ git_drop_cmd, shell=True)
             self.default_settings.erase('stashes')
+
+class TabsetgitCommand(sublime_plugin.WindowCommand):
+
+    def run(self):
+        self.window.show_quick_panel(['No', 'Yes'], self.save_git_setting)
+
+    def save_git_setting(self, option):
+        default_settings = sublime.load_settings("Tabstasher.sublime-settings")
+        default_settings.set('git',option)
